refactor: moved job to core
