+++ packages/core/src/index.ts
@@ -1,4 +1,5 @@
 export * from "./backends";
+export * from "./job";
 export * from "./logger";
 export * from "./schema";
 export * from "./tools";
