+++ packages/engine/src/shared-runner/runner-pool.ts
@@ -22,6 +22,7 @@ export class RunnerPool {
       filename: runnerPath,
       minThreads: this.nonNullConfig.minThreads,
       maxThreads: this.nonNullConfig.maxThreads,
+      idleTimeout: 10_000,
     });
     logger("RunnerPool").debug(
       `Created worker pool with min ${this.nonNullConfig.minThreads} threads and max ${this.nonNullConfig.maxThreads} threads`,
